+++ packages/uploader-server/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.1.92 (March 26, 2025)
+
+- feat: add initDynamicResourceMiddleware middleware
+
+## 0.1.91 (March 25, 2025)
+
+- refactor: polish initStaticResourceMiddleware logic
+
 ## 0.1.90 (March 16, 2025)
 
 - chore: update deps